Moved stringifyError into its own file as it was causing strict compile issues in ErrorHandlingUtils
